+++ MatrixSDK/VoIP/MXCall.m
@@ -96,7 +96,7 @@ NSString *const kMXCallStateDidChange = @"kMXCallStateDidChange";
         _endReason = MXCallEndReasonUnknown;
 
         // Consider we are using a conference call when there are more than 2 users
-        _isConferenceCall = (2 < _room.state.members.joinedMembers.count);
+        _isConferenceCall = (2 < _room.state.membersCount.joined);
         
         // Set caleeId only for regular calls
         if (!_isConferenceCall)