+++ MatrixSDKTests/MXPeekingRoomTests.m
@@ -73,7 +73,7 @@
                     XCTAssertEqual(mxSession.rooms.count, 0, @"MXPeekingRoom must not be listed by mxSession.rooms");
                     XCTAssertEqual(peekingRoom.roomId, room.roomId);
 
-                    XCTAssertEqual(peekingRoom.state.members.members.count, 1, @"The MXPeekingRoom state must be known now");
+                    XCTAssertEqual(peekingRoom.state.membersCount.members, 1, @"The MXPeekingRoom state must be known now");
 
                     [mxSession stopPeeking:peekingRoom];
 
@@ -126,15 +126,18 @@
 
         mxSession = mxSession2;
 
+        NSLog(@"___ %@", @(room.state.membersCount.members));
+        NSLog(@"___ %@", @(room.state.members.members.count));
+
         XCTAssertEqual(mxSession.rooms.count, 1);
-        XCTAssertEqual(room.state.members.members.count, 1);
+        XCTAssertEqual(room.state.membersCount.members, 1);
 
         [mxSession peekInRoomWithRoomId:room.roomId success:^(MXPeekingRoom *peekingRoom) {
 
             XCTAssertEqual(mxSession.rooms.count, 1, @"MXPeekingRoom must not be listed by mxSession.rooms");
             XCTAssertEqual(peekingRoom.roomId, room.roomId);
 
-            XCTAssertEqual(peekingRoom.state.members.members.count, 1, @"The MXPeekingRoom state must be known now");
+            XCTAssertEqual(peekingRoom.state.membersCount.members, 1, @"The MXPeekingRoom state must be known now");
 
             [mxSession stopPeeking:peekingRoom];