+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes in Matrix iOS SDK in 0.11.0 ()
 
 Improvements:
 * MXRestClient: Add Matrix filter API
+* MXRoomState: Add a membersCount struct to store members stats independently from MXRoomMember objects.
 
 Bug fix: